feat(mods): add generalStore module
This commit adds the generalStore module, which provides functions for managing a general store of key-value pairs. The module includes functions for showing the general store, changing the value of a key, setting a new key in the general store, and getting the value of a key from the general store.

The module is located in "./src/mods/generalStore.js".