feat(dependencieStore): add prefix support to store and retrieve values
This commit adds support for prefixes in the dependencieStore module. Now, when setting a key-value pair, you can specify a prefix to group related values. When retrieving a value, you can also provide a prefix to narrow down the search. This allows for better organization and retrieval of dependencies.

Changes:
- Added prefix parameter to setDependencieStoreKey function
- Modified changeValue function to use the prefix parameter
- Modified getDependencieValue function to handle prefix and key parameters

Fixes #123